packages/pytest-httpchain-userfunc/pyproject.toml
[project]
name = "pytest-httpchain-userfunc"
version = "0.2.2"
description = "User functions support for pytest-httpchain"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Alexander Eresov", email = "aeresov@gmail.com" }]
dependencies = ["pytest-httpchain-core"]

[build-system]
requires = ["uv_build>=0.7.21,<0.8.0"]
build-backend = "uv_build"
